fixing var types

+++ includes/socket.h
@@ -39,13 +39,13 @@ typedef struct socket_store_t {
 
     // This is the sender portion.
     uint8_t sendBuff[SOCKET_BUFFER_SIZE];
-    uint16_t lastWritten;
+    uint8_t lastWritten;
     uint16_t lastAck;
     uint16_t lastSent;
 
     // This is the receiver portion
     uint8_t rcvdBuff[SOCKET_BUFFER_SIZE];
-    uint16_t lastRead;
+    uint8_t lastRead;
     uint16_t lastRcvd;
     uint16_t nextExpected;
